+++ lucene/sandbox/src/java/org/apache/lucene/sandbox/codecs/idversion/SingleDocsEnum.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.lucene.codecs.idversion;
+package org.apache.lucene.sandbox.codecs.idversion;
 
 import java.io.IOException;
