+++ lucene/sandbox/src/resources/META-INF/services/org.apache.lucene.codecs.PostingsFormat
@@ -13,4 +13,4 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-org.apache.lucene.codecs.idversion.IDVersionPostingsFormat
+org.apache.lucene.sandbox.codecs.idversion.IDVersionPostingsFormat
